Copy over draft information on an unpublished private review request discard
If you create a new review request, fill out the description, testing done,
etc. information, and then discard it without publishing, the information that
you filled out will no longer be in the review request. With this fix, the
draft information is copied over on private unpublished review requests.

Testing Done:
- All existing unit tests were run and passed.
- Added two unit tests to check behaviour of discarding a private unpublished
  review request (this one should have the change description updated based on
  the request draft) and discarding a public unpublished review request (this
  should not have the change description updated and behaves like it used to).
- Manually tested combinations of: filling out the information (description,
  testing done, file attachments) of a new review request, discarding it,
  reopening it, modifying/not modifying fields, and discarding it again.

Bugs closed: 3465

Reviewed at https://reviews.reviewboard.org/r/6748/

+++ reviewboard/reviews/models/review_request.py
@@ -678,11 +678,19 @@ class ReviewRequest(BaseReviewRequestDetails):
         if type not in [self.SUBMITTED, self.DISCARDED]:
             raise AttributeError("%s is not a valid close type" % type)
 
+        draft = get_object_or_none(self.draft)
+
         if self.status != type:
-            # TODO: Use the user's default for rich_text.
-            changedesc = ChangeDescription(public=True,
-                                           text=description or "",
-                                           rich_text=rich_text or False)
+            if (draft is not None and
+                not self.public and type == self.DISCARDED):
+                # Copy over the draft information if this is a private discard.
+                changedesc = draft.publish(self, send_notification=False)
+            else:
+                # TODO: Use the user's default for rich_text.
+                changedesc = ChangeDescription(public=True,
+                                               text=description or "",
+                                               rich_text=rich_text or False)
+
             changedesc.record_field_change('status', self.status, type)
             changedesc.save()
 
@@ -710,11 +718,8 @@ class ReviewRequest(BaseReviewRequestDetails):
             # Needed to renew last-update.
             self.save()
 
-        try:
-            draft = self.draft.get()
-        except ObjectDoesNotExist:
-            pass
-        else:
+        # Delete the associated draft review request.
+        if draft is not None:
             draft.delete()
 
     def reopen(self, user=None):
